Add user roles support

+++ wfm/directory.php
@@ -59,8 +59,12 @@
 
                 if (isset($_SESSION['wfm_user']) && $file != '..') {
                     echo('<td class="td-admin">');
-                    echo('<a onclick="deleteFile(\'' . $file . '\', \'' . $fileName . '\');" href="#" class="admin-action-delete"><span class="' . getFAClass('delete') . '"></span></a>');
-                    echo('<a onclick="alert(\'Comming soon :)\');" href="#" class="admin-action-move" aria-disabled="true"><span class="' . getFAClass('move') . '"></span></a>');
+                    if (canUserDo($_SESSION['wfm_user'], 'delete')) {
+                        echo('<a onclick="deleteFile(\'' . $file . '\', \'' . $fileName . '\');" href="#" class="admin-action-delete"><span class="' . getFAClass('delete') . '"></span></a>');
+                    }
+                    if (canUserDo($_SESSION['wfm_user'], 'move')) {
+                        echo('<a onclick="alert(\'Comming soon :)\');" href="#" class="admin-action-move" aria-disabled="true"><span class="' . getFAClass('move') . '"></span></a>');
+                    }
                     echo('</td>');
                 }
 
@@ -71,35 +75,44 @@
 </table>
 
 <?php if (isset($_SESSION['wfm_user'])): ?>
-<form class="admin-action-upload" method="post" action="index.php" enctype="multipart/form-data">
-    <input type="hidden" name="admin" value="upload">
-    <input type="hidden" name="path" value="<?= PATH; ?>">
-    <label>Upload file:</label>
-    <input type="file" name="file" required>
-    <button type="submit" class="btn btn-success">
-        <span class="<?= getFAClass('upload'); ?>"></span>
-        Upload
-    </button>
-    <a onclick="createDir();" href="#" class="btn btn-warning">
-        <span class="<?= getFAClass('new'); ?>"></span>
-        Create new directory
-    </a>
-</form>
+    <form class="admin-action-upload" method="post" action="index.php" enctype="multipart/form-data">
+        <?php if (canUserDo($_SESSION['wfm_user'], 'upload')): ?>
+            <input type="hidden" name="admin" value="upload">
+            <input type="hidden" name="path" value="<?= PATH; ?>">
+            <label>Upload file:</label>
+            <input type="file" name="file" required>
+            <button type="submit" class="btn btn-success">
+                <span class="<?= getFAClass('upload'); ?>"></span>
+                Upload
+            </button>
+        <?php endif; ?>
 
-<script>
-    function deleteFile(file, name) {
-        if (confirm('Do you really want to delete ' + name + ' file (directory)?')) {
-            window.location.href = 'index.php?a=delete&d=<?= PATH; ?>&f=' + file;
-        }
-    }
+        <?php if (canUserDo($_SESSION['wfm_user'], 'dirs')): ?>
+            <a onclick="createDir();" href="#" class="btn btn-warning">
+                <span class="<?= getFAClass('new'); ?>"></span>
+                Create new directory
+            </a>
+        <?php endif; ?>
+    </form>
 
-    function createDir() {
-        var name = prompt('Directory name:', '');
-        if (name != null && name != "") {
-            window.location.href = 'index.php?a=create&d=<?= PATH; ?>&n=' + name;
-        }
-    }
-</script>
+    <script>
+        <?php if (canUserDo($_SESSION['wfm_user'], 'delete')): ?>
+            function deleteFile(file, name) {
+                if (confirm('Do you really want to delete ' + name + ' file (directory)?')) {
+                    window.location.href = 'index.php?a=delete&d=<?= PATH; ?>&f=' + file;
+                }
+            }
+        <?php endif; ?>
+
+        <?php if (canUserDo($_SESSION['wfm_user'], 'dirs')): ?>
+            function createDir() {
+                var name = prompt('Directory name:', '');
+                if (name != null && name != "") {
+                    window.location.href = 'index.php?a=create&d=<?= PATH; ?>&n=' + name;
+                }
+            }
+        <?php endif; ?>
+    </script>
 <?php endif; ?>
 
 <?php
